feat(test-utils): restructure mock store handlers and update exports
- Added new handlers for file tree, files, and versions.
- Updated `setupMockStore` to utilize new handlers.
- Changed exports in `index.ts` and `package.json` to reflect new structure.
- Removed deprecated `store.ts` and related definitions.
