+++ packages/test-utils/src/mock-store/handlers/file-tree.ts
@@ -1,16 +1,20 @@
-import { HttpResponse } from "../msw";
-import { defineMockFetchHandler } from "./__define";
+import type { HandlerContext } from "../types";
+import { HttpResponse, mockFetch } from "../../msw";
 
-export const fileTreeMockHandler = defineMockFetchHandler("/api/v1/versions/:version/file-tree", ({ baseUrl, response }) => {
+export function setupFileTreeHandler({
+  baseUrl,
+  response
+}: HandlerContext<"/api/v1/versions/:version/file-tree">): void {
   const url = `${baseUrl}/api/v1/versions/:version/file-tree`;
 
   if (typeof response === "function") {
-    return [
+    mockFetch([
       ["GET", url, response],
-    ];
+    ]);
+    return;
   }
 
-  return [
+  mockFetch([
     ["GET", url, () => {
       if (response === true || response == null) {
         return HttpResponse.json([
@@ -45,5 +49,5 @@ export const fileTreeMockHandler = defineMockFetchHandler("/api/v1/versions/:ver
 
       return HttpResponse.json(response);
     }],
-  ];
-});
+  ]);
+}
