+++ packages/test-utils/tsdown.config.ts
@@ -4,7 +4,7 @@ export default createTsdownConfig({
   entry: {
     "index": "./src/index.ts",
     "msw": "./src/msw.ts",
-    "store": "./src/store.ts",
-    "vitest-setup": "./src/msw/global-setup.ts",
+    "mock-store": "./src/mock-store/index.ts",
+    "msw/vitest-setup": "./src/msw/vitest-setup.ts",
   },
 });
